==> 01-getting-started.Rmd <==

# Getting started with R {#getting-started}

There is a lot of great material out there to get you started with R. I enjoyed [swirl](https://swirlstats.com), which teaches basic R, in R. Whatever you choose, it is probably a good idea to familiarize yourself to some extent with basic R. Below, I only treat some of the basic stuff that is needed to run the code that is presented in the other chapters.

## Running R

R is free and it is available for many different platforms. You can run [plain R](https://www.r-project.org), run it from [RStudio Desktop](https://www.rstudio.com) or even from a [browser](https://rdrr.io/snippets/). I prefer Rstudio and the instructions are based on Rstudio as well.
When you start Rstudio, you see a couple of windows, including a 'Console'. This is the place where you can type your commands for R.

## Using the command line

A classic in coding language is to print 'hello world'. To do this, you need to type this piece of code (followed by the enter key) in the console:

`print('hello world')`

The result, printed below the code in the console is:

`[1] "hello world"` 

In this document the input and output is shown as grey boxes. The first grey box represents the input and, if there is any output, the second grey box shows the output, for example:

```{r}
print('hello world')
```

It is possible to copy the code from the first box. When you move your cursor to the upper left corner of the box a copy icon will appear. If you click on the icon, the code is copied.

Not every piece of code results in a visible output, for instance when I assign a value to the variable x:

```{r}
x <- 1
```

To show the value of a variable as output, type its name:

```{r}
x
```

R comes with datasets. Although these datasets are not so relevant for us, they are often used to demonstrate functions in R. One of these datasets is `mtcars` and we can use the `head()` function to check the first lines:

```{r}
head(mtcars)
```

## ?Help

If you need help with a specific function, you may look at the documentation by typing a question mark followed by the function:

```{r}
?head()
```

However, I usually find myself googling instead of reading the documentation. The result is often a website where the correct use is explained by an example. It also make sense to experiment. Systematically changing variables or options (or outcommenting lines of code, see below) will teach you the logic behind a function or operation. Finally, there may be some answers in the chapter [Questions and Answers](#questions-and-answers)

## Installing packages

R comes with a lot of commands and functions, but we often need to load additional 'packages' to add functionality. The most important one that we use here is the `{ggplot2}` package that we use for the data visualizations. The package can be activated with the function `require()` or `library()`:

```{r}
library("ggplot2")
```

The `{tidyverse}` package is very versatile and is a superpackage that hold several packages (including `{ggplot2}`). Loading the `{tidyverse}` package is like equipping R with superpowers. Often it is sufficient to load the `{tidyverse}` package. 

The first time, the package has to be downloaded and installed. This handy piece of code checks whether the `{tidyverse}` package is available, downloads it if necessary and than activates it:

```{r, message=FALSE}
if (!require(tidyverse)) {
    install.packages("tidyverse")
    require(tidyverse)
}
```

The output in the console depends on the packages that are installed and activated. For some specific functions, other packages are required, and these will be mentioned whenever they are used.

## Multiline code

The tidyverse package introduces a so-called pipe operator `%>%` which we will use a lot. This operator is useful for stringing multiple functions together. An example is given below, which reads as 'take the `mtcars` dataset _and next use_ the `head()` function'.

```{r}
mtcars %>% head()
```

Also for the plots that are generated with `ggplot()`, several functions can be added and these reflects different layers in the plot. However, in case of ggplot the layers are combined by a `+`:
```{r}
ggplot(mtcars, aes(wt, mpg)) +
  geom_point() +
  geom_line()
```

It is important to end each line with a `+`, as it indicates that the code continues on the next line. This will result in a warning message:

```{r, eval=FALSE}
ggplot(mtcars, aes(wt, mpg)) +
  geom_point()
  + geom_line()
```

The last line should not end with a plus. A trick that I use a lot, is using `NULL` on the last line of the code:

```{r}
ggplot(mtcars, aes(wt, mpg)) +
  geom_point() +
  geom_line() +
  NULL
```

The advantage is that it is easy to deactivate a line by placing a hashtag in front of it (without the need to remove the `+`). The hashtag tells R to treat that line as a comment and not as code:

```{r}
ggplot(mtcars, aes(wt, mpg)) +
  geom_point() +
# geom_line() +
  NULL
```

This strategy is called 'commenting out' and is very useful to examine the effect of a line of code. For plots, it works very well in combination with `NULL` on the last line.
